+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2.0 branch
 ### 2.0.1-beta
 * fixed bug, now the cache is flushed automatically, if there's a post-dated post to be published;
-* fixed a bug with the posts date.
+* fixed bug with the posts date.
 
 ### 2.0.0-alpha
 * all the code has been completely rewritten for CakePHP 3.x. Several optimizations have been applied;