+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fixed bug, now the cache is flushed automatically, if there's a post-dated post to be published;
 * fixed permissions;
 * it automatically adds the meta tag for RSS resources;
+* the backend menus are generated fully automatically. You no longer need any configuration;
 * now you can choise which post details display using the configuration file;
 * now you can set the timezone using the configuration file;
 * fixed bug with the posts date.