+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.x branch
 ## 2.0 branch
-### 2.0.1-beta
+### 2.0.1-RC1
 * fixed bug, now the cache is flushed automatically, if there's a post-dated post to be published;
 * fixed bug on login with cookie;
 * fixed permissions;